fix: PR レビュー指摘対応 — 依存方向修正と invalid アクション追加
- PedigreeRow を pedigree/service.ts から horses/types.ts に移動し
  Repository → Service の逆依存を解消
- ImportPreviewRow に action: 'invalid' と skipReason を追加し
  入力不備（birthYear null）と変更なしスキップを明確に区別
- ImportPreview.summary に invalidCount を追加
